backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
google-generativeai==0.8.3
openai==1.50.0
Pillow==10.4.0
reportlab==4.1.0
aiofiles==24.1.0
httpx==0.27.0
tenacity==9.0.0
